refactor: payment 패키지 정의 및 Paymentclient 위치 이동
